Remove command-specific code from the lib
